app/backend/app/db.py
import psycopg2
from app.config import settings

def check_db_connection():
    conn = psycopg2.connect(settings.DATABASE_URL)
    cur = conn.cursor()
    cur.execute("SELECT 1;")
    cur.close()
    conn.close()
    return True